Partial fix of #36.
I have added links from a utilityNetwork to all its elements as a start
of solving issue #36. The rest of the links that should be added I
haven’t done yet. Mostly because I want to discuss with @martinborgman
and @PalmJanssen whether the alternative (removing the association name
from the model) is a viable alternative.